services/ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
gunicorn==21.2.0
pyyaml==6.0.1
